More Lint

+++ src/main/java/net/kyrptonaught/customportalapi/util/ColorUtil.java
@@ -10,7 +10,7 @@ public record ColorUtil() {
             int l = (color & 16711680) >> 16;
             int m = (color & '\uff00') >> 8;
             int n = (color & 255);
-            COLOR_CACHE.put(color, new float[]{(float) l / 255.0F, (float) m / 255.0F, (float) n / 255.0F});
+            COLOR_CACHE.put(color, new float[]{l / 255.0F, m / 255.0F, n / 255.0F});
         }
         return COLOR_CACHE.get(color);
     }
